+++ scripts/test-regressions.sh
@@ -8,6 +8,9 @@ trap 'rm -rf "$OUT"' EXIT
 
 xcrun swiftc -O \
   DropDrive/Utilities/DestinationCapacity.swift \
+  DropDrive/Utilities/GoogleDriveLinkParser.swift \
+  DropDrive/Utilities/LinkIdentity.swift \
+  DropDrive/Utilities/SupportedLinkExtractor.swift \
   DropDrive/Models/ExternalLinkReceipt.swift \
   DropDrive/Utilities/UniqueDestinationNaming.swift \
   DropDrive/Utilities/SecurityScopedAccessManager.swift \